+++ aoc2019/day_three/wire_runner.py
@@ -1,7 +1,10 @@
 import sys
 from _collections import defaultdict
 
-"""This module provides the logic for day 3 aoc2019 solution with utility functions and classes"""
+"""
+This module provides the logic for day 3 aoc2019 solution with utility
+functions and classes.
+"""
 
 
 def init_grid(size):
@@ -10,12 +13,14 @@ def init_grid(size):
     of lists. Always makes a square grid with odd dimensions
     """
     # creates a list of lists with unique elements valued at 0
-    return [[Empty() for j in range(size)] for i in range(size + (size + 1) % 2)]
+    odd_size = size + (size + 1) % 2
+    return [[Empty() for j in range(odd_size)] for i in range(odd_size)]
 
 
 def resize_grid(grid):
     """
-    Utility function to make a new grid double in size and transfer existing entities
+    Utility function to make a new grid double in size and transfer existing
+    entities.
     """
     # Initialize a grid roughly
     new_size = 2 * len(grid) + 1
@@ -23,7 +28,7 @@ def resize_grid(grid):
 
     # iteratively copy grid into new larger grid
     offset = int((new_size - len(grid)) / 2)
-    for i in range (len(grid)):
+    for i in range(len(grid)):
         for j in range(len(grid)):
             new_grid[i + offset][j + offset] = grid[i][j]
 
@@ -32,8 +37,8 @@ def resize_grid(grid):
 
 def get_intersections(grid):
     """
-    This utility function finds all the intersections in a grid. It returns a list
-    of (x, y) tuples
+    This utility function finds all the intersections in a grid. It returns a
+    list of (x, y) tuples.
     """
     cross_list = []  # initialize empty list
     # Iterate through all cells in the grid - find intersections
@@ -66,9 +71,10 @@ def get_min_dist_origin(origin, intersections):
 
 class WireGrid:
     """
-    Class to represent a grid with Wires mapped on it. Provides attributes and behaviors
-    to map "wires" onto the grid based a list of "tracing" instructions. Works when grid is small. Becomes
-    unmanageable with large grid sizes.
+    Class to represent a grid with Wires mapped on it. Provides attributes and
+    behaviors to map "wires" onto the grid based a list of "tracing"
+    instructions. Works when grid is small. Becomes unmanageable with large
+    grid sizes.
     """
     INITIAL_SIZE = 1021  # Constant defining initial grid size
 
@@ -82,21 +88,27 @@ class WireGrid:
         # Dictionary storing current "position"
         self.current_pos = {'x': self.center[0], 'y': self.center[1]}
         # Map instruction codes to helper methods
-        self.instructions = {'U': self.up, 'D': self.down, 'R': self.right, 'L': self.left}
+        self.instructions = {'U': self.up,
+                             'D': self.down,
+                             'R': self.right,
+                             'L': self.left}
         # Map actions based on cell content
         self.actions = {' ': Wire, '.': Cross}
 
     def get_min_manhattan_dist(self):
-        """This method returns the manhattan distance from closest intersection from origin"""
+        """
+        This method returns the manhattan distance from closest intersection
+        from the origin.
+        """
         intersections = get_intersections(self.grid)
         dist = get_min_dist_origin(self.center, intersections)
         return dist
 
     def trace_wires(self, trace):
         """
-        Method maps a series of passed "trace" instructions to the grid. Expected
-        instructions are in the form of D123 to go down 123 cells or R03 to go right
-        3 cells.
+        Method maps a series of passed "trace" instructions to the grid.
+        Expected instructions are in the form of D123 to go down 123 cells
+        or R03 to go right 3 cells.
         """
         # reset "current" position at center
         self.current_pos = {'x': self.center[0], 'y': self.center[1]}
@@ -104,23 +116,21 @@ class WireGrid:
         self.grid[self.center[1]][self.center[0]] = End()
         # iterate through "trace" list
         for opp in trace:
-            #try:
             self.instructions[opp[0]](int(opp[1:]))
-            # grid not large enough
-            #except IndexError:
-            #    self.grid = resize_grid(self.grid)  # enlarge grid
-            #    self.size = len(self.grid)  # memoization - store size
-            #    self.center = (int(self.size / 2), int(self.size / 2)) # recenter grid
-            #    self.instructions[opp[0]](int(opp[1:]))  # try again with "trace" step
 
     def up(self, dist):
-        """Helper method to trace "up" a passed number of cells from current position"""
+        """
+        Helper method to trace "up" a passed number of cells from current
+        position.
+        """
         # iterate up
         for offset in range(1, dist + 1):
             try:
-                current = self.grid[self.current_pos['y'] - offset][self.current_pos['x']]
+                current = self.grid[self.current_pos['y']
+                                    - offset][self.current_pos['x']]
                 # dynamically dispatch action
-                self.grid[self.current_pos['y'] - offset][self.current_pos['x']] = \
+                self.grid[self.current_pos['y']
+                          - offset][self.current_pos['x']] = \
                     self.actions[current.__str__()]()
             except IndexError:
                 pass  # do nothing
@@ -128,13 +138,18 @@ class WireGrid:
         self.current_pos['y'] -= dist
 
     def down(self, dist):
-        """Helper method to trace "down" a passed number of cells from current position"""
+        """
+        Helper method to trace "down" a passed number of cells from current
+        position.
+        """
         # iterate down
         for offset in range(1, dist + 1):
             try:
-                current = self.grid[self.current_pos['y'] + offset][self.current_pos['x']]
+                current = self.grid[self.current_pos['y']
+                                    + offset][self.current_pos['x']]
                 # dynamically dispatch action
-                self.grid[self.current_pos['y'] + offset][self.current_pos['x']] = \
+                self.grid[self.current_pos['y']
+                          + offset][self.current_pos['x']] = \
                     self.actions[current.__str__()]()
             except IndexError:
                 pass  # do nothing
@@ -142,28 +157,42 @@ class WireGrid:
         self.current_pos['y'] += dist
 
     def right(self, dist):
-        """Helper method to trace "right" a passed number of cells from current position"""
+        """
+        Helper method to trace "right" a passed number of cells from current
+        position.
+        """
         # iterate up
         for offset in range(1, dist + 1):
             try:
-                current = self.grid[self.current_pos['y']][self.current_pos['x'] + offset]
+                current = \
+                    self.grid[self.current_pos['y']][self.current_pos['x']
+                                                     + offset]
                 # dynamically dispatch action
-                self.grid[self.current_pos['y']][self.current_pos['x'] + offset] = \
-                    self.actions[current.__str__()]()
+                current_y = self.current_pos['y']
+                offset_x = self.current_pos['x'] + offset
+                self.grid[current_y][offset_x] \
+                    = self.actions[current.__str__()]()
             except IndexError:
                 pass  # do nothing
         # Update "current" coords
         self.current_pos['x'] += dist
 
     def left(self, dist):
-        """Helper method to trace "left" a passed number of cells from current position"""
+        """
+        Helper method to trace "left" a passed number of cells from current
+        position.
+        """
         # iterate up
         for offset in range(1, dist + 1):
             try:
-                current = self.grid[self.current_pos['y']][self.current_pos['x'] - offset]
+                current = \
+                    self.grid[self.current_pos['y']][self.current_pos['x']
+                                                     - offset]
                 # dynamically dispatch action
-                self.grid[self.current_pos['y']][self.current_pos['x'] - offset] = \
-                    self.actions[current.__str__()]()
+                current_y = self.current_pos['y']
+                offset_x = self.current_pos['x'] - offset
+                self.grid[current_y][offset_x] \
+                    = self.actions[current.__str__()]()
             except IndexError:
                 pass  # do nothing
         # Update "current" coords
@@ -197,7 +226,7 @@ class Wire:
         self.taken = True
 
     def __str__(self):
-       return '.'
+        return '.'
 
 
 class Cross:
@@ -225,18 +254,25 @@ class WireTable:
         self.current_steps = 0  # counter for steps
         self.current_trace = None  # dictionary of current trace
         self.intersections = None  # list of intersection tuples
-        self.list_keyed_intersect_steps = None  # list of dictionaries of steps per intersection
+        # list of dictionaries of steps per intersection
+        self.list_keyed_intersect_steps = None
         self.current_intersect_steps = None  # dictionary of current steps
         # Map instruction codes to helper methods
-        self.instructions = {'U': self.up, 'D': self.down, 'R': self.right, 'L': self.left}
+        self.instructions = {'U': self.up,
+                             'D': self.down,
+                             'R': self.right,
+                             'L': self.left}
         # Map instruction codes to helper methods
-        self.count_steps_instr = {'U': self.up_count, 'D': self.down_count, 'R': self.right_count, 'L': self.left_count}
+        self.count_steps_instr = {'U': self.up_count,
+                                  'D': self.down_count,
+                                  'R': self.right_count,
+                                  'L': self.left_count}
 
     def trace_wires(self, trace):
         """
-        Method maps a series of passed "trace" instructions to the trace dictionaries. Expected
-        instructions are in the form of D123 to go down 123 cells or R03 to go right
-        3 cells.
+        Method maps a series of passed "trace" instructions to the trace
+        dictionaries. Expected instructions are in the form of D123 to go
+        down 123 cells or R03 to go right 3 cells.
         """
         # save a reference to the list
         self.trace_list.append(trace)
@@ -254,7 +290,10 @@ class WireTable:
         self.current_trace = None
 
     def get_min_manhattan_dist(self):
-        """This method returns the manhattan distance from closest intersection from origin"""
+        """
+        This method returns the manhattan distance from closest intersection
+        from origin.
+        """
         # generate intersections
         self.find_intersections()
         # get manhattan distance to closest
@@ -262,7 +301,10 @@ class WireTable:
         return dist
 
     def find_intersections(self):
-        """Helper method that iteratively saves list of intersections for both traces"""
+        """
+        Helper method that iteratively saves list of intersections for both
+        traces.
+        """
         self.intersections = []
         # iterate through
         for x_trace1 in self.traces[0]:
@@ -277,6 +319,9 @@ class WireTable:
                 pass  # do nothing - no matching pairs at this x value
 
     def find_min_steps(self):
+        """
+        Find the minimum number of steps required to reach an intersection.
+        """
         self.list_keyed_intersect_steps = []
         for trace in self.trace_list:
             self.current_intersect_steps = {}
@@ -289,18 +334,23 @@ class WireTable:
             for opp in trace:
                 self.count_steps_instr[opp[0]](int(opp[1:]))
             # Append
-            self.list_keyed_intersect_steps.append(self.current_intersect_steps)
+            self.list_keyed_intersect_steps\
+                .append(self.current_intersect_steps)
 
         # calculate sum of steps to a given intersection
         min_steps = sys.maxsize
         for cross in self.intersections:
-            steps = self.list_keyed_intersect_steps[0][cross] + self.list_keyed_intersect_steps[1][cross]
+            steps = self.list_keyed_intersect_steps[0][cross] \
+                    + self.list_keyed_intersect_steps[1][cross]
             min_steps = min(steps, min_steps)
         # minimum number of steps to intersection calculated
         return min_steps
 
     def up(self, dist):
-        """Helper method to trace "up" a passed number of cells from current position"""
+        """
+        Helper method to trace "up" a passed number of cells from current
+        position.
+        """
         # iterate up
         for offset in range(1, dist + 1):
             self.current_trace[self.current_x].append(self.current_y + offset)
@@ -308,7 +358,10 @@ class WireTable:
         self.current_y += dist
 
     def down(self, dist):
-        """Helper method to trace "down" a passed number of cells from current position"""
+        """
+        Helper method to trace "down" a passed number of cells from current
+        position.
+        """
         # iterate down
         for offset in range(1, dist + 1):
             self.current_trace[self.current_x].append(self.current_y - offset)
@@ -316,7 +369,10 @@ class WireTable:
         self.current_y -= dist
 
     def right(self, dist):
-        """Helper method to trace "right" a passed number of cells from current position"""
+        """
+        Helper method to trace "right" a passed number of cells from current
+        position.
+        """
         # iterate right
         for offset in range(1, dist + 1):
             self.current_trace[self.current_x + offset].append(self.current_y)
@@ -324,7 +380,10 @@ class WireTable:
         self.current_x += dist
 
     def left(self, dist):
-        """Helper method to trace "left" a passed number of cells from current position"""
+        """
+        Helper method to trace "left" a passed number of cells from current
+        position.
+        """
         # iterate left
         for offset in range(1, dist + 1):
             self.current_trace[self.current_x - offset].append(self.current_y)
@@ -333,57 +392,65 @@ class WireTable:
 
     def up_count(self, dist):
         """
-        Helper method to trace "up" a passed number of cells from current position
-        and count steps to and intersection.
+        Helper method to trace "up" a passed number of cells from current
+        position and count steps to and intersection.
         """
         # iterate up
         for offset in range(1, dist + 1):
             self.current_steps += 1  # increment step counter
             # currently at
             if (self.current_x, self.current_y + offset) in self.intersections:
-                self.current_intersect_steps[(self.current_x, self.current_y + offset)] = self.current_steps
+                self.current_intersect_steps[(self.current_x,
+                                              self.current_y + offset)]\
+                    = self.current_steps
 
         # Update "current" coords
         self.current_y += dist
 
     def down_count(self, dist):
         """
-        Helper method to trace "down" a passed number of cells from current position
-        and count steps to and intersection.
+        Helper method to trace "down" a passed number of cells from current
+        position and count steps to and intersection.
         """
         # iterate down
         for offset in range(1, dist + 1):
             self.current_steps += 1  # increment step counter
             # currently at
             if (self.current_x, self.current_y - offset) in self.intersections:
-                self.current_intersect_steps[(self.current_x, self.current_y - offset)] = self.current_steps
+                self.current_intersect_steps[(self.current_x,
+                                              self.current_y - offset)] \
+                    = self.current_steps
         # Update "current" coords
         self.current_y -= dist
 
     def right_count(self, dist):
         """
-        Helper method to trace "right" a passed number of cells from current position
-        and count steps to and intersection.
+        Helper method to trace "right" a passed number of cells from current
+        position and count steps to and intersection.
         """
         # iterate right
         for offset in range(1, dist + 1):
             self.current_steps += 1  # increment step counter
             # currently at
             if (self.current_x + offset, self.current_y) in self.intersections:
-                self.current_intersect_steps[(self.current_x + offset, self.current_y)] = self.current_steps
+                self.current_intersect_steps[(self.current_x + offset,
+                                              self.current_y)] \
+                    = self.current_steps
         # Update "current" coords
         self.current_x += dist
 
     def left_count(self, dist):
         """
-        Helper method to trace "left" a passed number of cells from current position
-        and count steps to and intersection.
+        Helper method to trace "left" a passed number of cells from current
+        position and count steps to and intersection.
         """
         # iterate left
         for offset in range(1, dist + 1):
             self.current_steps += 1  # increment step counter
             # currently at
             if (self.current_x - offset, self.current_y) in self.intersections:
-                self.current_intersect_steps[(self.current_x - offset, self.current_y)] = self.current_steps
+                self.current_intersect_steps[(self.current_x - offset,
+                                              self.current_y)] \
+                    = self.current_steps
         # Update "current" coords
         self.current_x -= dist