+++ aoc2019/run_day_five.py
@@ -1,6 +1,9 @@
 from shared.int_code import IntCode
 
-"""Runner script for Day 5 Advent of Code challenge - relies on refactored IntCode class for logic"""
+"""
+Runner script for Day 5 Advent of Code challenge - relies on refactored
+IntCode class for logic
+"""
 
 # Get list of numbers from input
 input_file = open('day_five/input_day_five.txt', 'r')
@@ -18,12 +21,11 @@ d_code = int_code_p1.output_codes[-1]
 print(d_code)
 
 # Part Two - run diagnostic tests with jump and comparative instruction methods
-print('Part Two - run diagnostic tests with jump and comparator instruction methods: ')
+print('Part Two - run diagnostic tests with jump and comparator instruction '
+      'methods: ')
 # print(input_codes[:])
 int_code_p2 = IntCode(input_codes.copy())
 int_code_p2.run(5)
 d_code = int_code_p2.output_codes[-1]
 # expect 10428568
 print(d_code)
-
-